# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/password-reset/request": {
      "post": {
        "operationId": "AuthController_requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/password-reset/complete": {
      "post": {
        "operationId": "AuthController_completePasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetCompleteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/2fa/verify": {
      "post": {
        "operationId": "AuthController_verifyTwoFactor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFactorVerifyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/ai/actions/authorize": {
      "post": {
        "operationId": "AiController_authorize",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/providers": {
      "post": {
        "operationId": "AiController_createProvider",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiProviderCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/providers/{id}/test": {
      "post": {
        "operationId": "AiController_testProvider",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiProviderTestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ai"
        ]
      }
    },
...
    "/approvals": {
      "post": {
        "operationId": "ApprovalsController_request",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "approvals"
        ]
      },
      "get": {
        "operationId": "ApprovalsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "approvals"
        ]
      }
    },
    "/approvals/{id}/decision": {
      "post": {
        "operationId": "ApprovalsController_decide",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "approvals"
        ]
      }
    },
    "/approvals/refunds/{id}/approve": {
      "post": {
        "operationId": "ApprovalsController_approveRefund",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/dispatch/assign-driver": {
      "post": {
        "operationId": "DispatchController_assignDriver",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignDriverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "dispatch"
        ]
      }
    },
...
  },
  "info": {
    "title": "AI-Agentic Smart Logistics Operating System API",
    "description": "Industrial modular monolith API contract.",
    "version": "0.1.0-rc.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "twoFactorCode": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "PasswordResetRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "PasswordResetCompleteDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "TwoFactorVerifyDto": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "challengeId",
          "code"
        ]
      },
...
      "AssignDriverDto": {
        "type": "object",
        "properties": {
          "shipmentId": {
            "type": "string"
          },
          "driverId": {
            "type": "string"
          },
          "packageStatus": {
            "type": "string",
            "enum": [
              "created",
              "reserved",
              "picking",
              "scanned",
              "packed",
              "label_generated",
              "staged",
              "ready_for_dispatch",
              "dispatched",
              "delivered",
              "returned",
              "quarantined",
              "damaged",
              "lost"
            ],
            "default": "ready_for_dispatch"
          }
        },
        "required": [
          "shipmentId",
          "driverId"
        ]
      },
...
      "AiActionDto": {
        "type": "object",
        "properties": {
          "agentCode": {
            "type": "string"
          },
          "requestedAction": {
            "type": "string"
          },
          "riskLevel": {
            "type": "string",
            "enum": [
              "L1_LOW",
              "L2_MEDIUM",
              "L3_HIGH",
              "L4_CRITICAL",
              "L5_PROHIBITED"
            ]
          },
          "actorRoles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "customer",
                "merchant",
                "shipper",
                "warehouse_staff",
                "warehouse_manager",
                "driver",
                "fleet_manager",
                "carrier",
                "freight_forwarder",
                "logistic_disponent",
                "support_agent",
                "finance_admin",
                "compliance_admin",
                "super_admin",
                "ai_agent"
              ]
            }
          },
          "approvalCount": {
            "type": "number"
          },
          "l2AutoPolicyEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "agentCode",
          "requestedAction",
          "actorRoles"
        ]
      },
      "AiProviderCreateDto": {
        "type": "object",
        "properties": {
          "apiKey": {
            "type": "string"
          },
          "providerName": {
            "type": "string"
          }
        },
        "required": [
          "apiKey",
          "providerName"
        ]
      },
      "AiProviderTestDto": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string"
          },
          "prompt": {
            "type": "string"
          }
        }
      },
...
      "ApprovalRequestDto": {
        "type": "object",
        "properties": {
          "actionCode": {
            "type": "string"
          },
          "riskLevel": {
            "type": "string",
            "enum": [
              "L1_LOW",
              "L2_MEDIUM",
              "L3_HIGH",
              "L4_CRITICAL",
              "L5_PROHIBITED"
            ]
          },
          "context": {
            "type": "object"
          }
        },
        "required": [
          "actionCode",
          "context"
        ]
      },
      "ApprovalDecisionDto": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "approved",
              "rejected"
            ]
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "decision"
        ]
      },
      "CommentDto": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string"
          }
        }
      }
    }
  }
}